navigation robust to urls changes, reorder components
