+++ browser/src/components/App.vue
@@ -43,8 +43,16 @@
 
 <script lang="ts">
 import SidebarMain from './sidebar/Sidebar.vue'
-import { mapState, mapMutations } from 'vuex'
+import TaskList from './pages/TaskList.vue'
+import MainTaskFrame from './navigation/Navigation.vue'
+import MainDescriptionFrame from './navigation/steps/DescriptionStep.vue'
+import MainTrainingFrame from './navigation/steps/TrainingStep.vue'
+import MainTestingFrame from './navigation/steps/TestingStep.vue'
 import { Platform } from '../platforms/platform'
+import { loadTasks } from '../tasks'
+
+import { Task } from 'discojs'
+import { mapState, mapMutations } from 'vuex'
 import { defineComponent } from 'vue'
 
 export default defineComponent({
@@ -75,8 +83,55 @@ export default defineComponent({
        */
     this.initPlatform(this.getPlatform())
   },
+  async created () {
+    // Load tasks and create dynamic routes
+    this.routesComponents = [
+      ['description', MainDescriptionFrame],
+      ['training', MainTrainingFrame],
+      ['testing', MainTestingFrame]
+    ]
+    const tasks: Task[] = await loadTasks()
+    tasks.forEach((task: Task) => {
+      const route = `/${task.taskID}`
+      this.$router.addRoute({
+        path: route,
+        // So that KeepAlive can differentiate the components
+        component: defineComponent({
+          key: task.taskID,
+          extends: MainTaskFrame
+        }),
+        props: {
+          id: task.taskID,
+          task: task
+        },
+        children: this.getTaskRoutes(task)
+      })
+    })
+    this.$router.addRoute({
+      path: '/list',
+      component: TaskList,
+      props: { tasks: tasks }
+    })
+  },
   methods: {
     ...mapMutations(['setIndexedDB', 'setAppTheme', 'setPlatform']),
+    getTaskRoutes (task: Task) {
+      return this.routesComponents.map((m) => {
+        const [route, component] = m
+        return {
+          path: route,
+          // So that KeepAlive can differentiate the components
+          component: defineComponent({
+            key: `${route}:${task.taskID}`,
+            extends: component
+          }),
+          props: {
+            id: task.taskID,
+            task: task
+          }
+        }
+      })
+    },
     getBrowserTheme () {
       if (window.localStorage.getItem('dark')) {
         return JSON.parse(window.localStorage.getItem('dark'))
