+++ browser/src/components/ModelActorFrame.vue
@@ -41,9 +41,9 @@
 </template>
 
 <script>
-import IconCard from '../../containers/IconCard.vue'
-import CheckList from '../../../assets/svg/CheckList.vue'
-import FileEarmarkRuled from '../../../assets/svg/FileEarmarkRuled.vue'
+import IconCard from './containers/IconCard.vue'
+import CheckList from '../assets/svg/CheckList.vue'
+import FileEarmarkRuled from '../assets/svg/FileEarmarkRuled.vue'
 import { Task } from 'discojs'
 
 export default {
