+++ browser/src/components/dataset_input/FileSelectionFrame.vue
