+++ browser/src/components/dataset_input/preview/PreviewGallery.vue
