+++ browser/src/components/dataset_input/preview/data_types/ImageData.vue
