+++ browser/src/components/navigation/ProgressBar.vue
