+++ browser/src/components/pages/Home.vue
@@ -99,21 +99,12 @@
 </template>
 
 <script lang="ts">
-import BaseLayout from './containers/BaseLayout.vue'
-import TitleCard from './containers/TitleCard.vue'
-import CustomHeader from './simple/CustomHeader.vue'
-import CustomButton from './simple/CustomButton.vue'
-import FederatedImage from '../assets/svg/FederatedImage.vue'
-import DecentralizedImage from '../assets/svg/DecentralizedImage.vue'
-import MainTaskFrame from './main_frames/MainTaskFrame.vue'
-import TaskList from './TaskList.vue'
-import MainDescriptionFrame from './main_frames/MainDescriptionFrame.vue'
-import MainTrainingFrame from './main_frames/MainTrainingFrame.vue'
-import MainTestingFrame from './main_frames/MainTestingFrame.vue'
-import { loadTasks } from '../tasks'
-
-import { Task } from 'discojs'
-import { defineComponent } from 'vue'
+import BaseLayout from '../containers/BaseLayout.vue'
+import TitleCard from '../containers/TitleCard.vue'
+import CustomHeader from '../simple/CustomHeader.vue'
+import CustomButton from '../simple/CustomButton.vue'
+import FederatedImage from '../../assets/svg/FederatedImage.vue'
+import DecentralizedImage from '../../assets/svg/DecentralizedImage.vue'
 
 export default {
   name: 'HomePage',
@@ -125,53 +116,7 @@ export default {
     DecentralizedImage,
     CustomHeader
   },
-  async created () {
-    this.routesComponents = [
-      ['description', MainDescriptionFrame],
-      ['training', MainTrainingFrame],
-      ['testing', MainTestingFrame]
-    ]
-    const tasks: Task[] = await loadTasks()
-    tasks.forEach((task: Task) => {
-      const route = `/${task.taskID}`
-      this.$router.addRoute({
-        path: route,
-        // So that KeepAlive can differentiate the components
-        component: defineComponent({
-          key: task.taskID,
-          extends: MainTaskFrame
-        }),
-        props: {
-          id: task.taskID,
-          task: task
-        },
-        children: this.getTaskRoutes(task)
-      })
-    })
-    this.$router.addRoute({
-      path: '/list',
-      component: TaskList,
-      props: { tasks: tasks }
-    })
-  },
   methods: {
-    getTaskRoutes (task: Task) {
-      return this.routesComponents.map((m) => {
-        const [route, component] = m
-        return {
-          path: route,
-          // So that KeepAlive can differentiate the components
-          component: defineComponent({
-            key: `${route}:${task.taskID}`,
-            extends: component
-          }),
-          props: {
-            id: task.taskID,
-            task: task
-          }
-        }
-      })
-    },
     goToTaskList () {
       this.$router.push({
         path: '/list'
