+++ browser/src/components/pages/NotFound.vue
@@ -45,10 +45,10 @@
 
 <script>
 import { defineComponent } from 'vue'
-import BaseLayout from './containers/BaseLayout.vue'
-import CustomButton from './simple/CustomButton.vue'
-import CustomHeader from './simple/CustomHeader.vue'
-import TitleCard from './containers/TitleCard.vue'
+import BaseLayout from '../containers/BaseLayout.vue'
+import CustomButton from '../simple/CustomButton.vue'
+import CustomHeader from '../simple/CustomHeader.vue'
+import TitleCard from '../containers/TitleCard.vue'
 
 export default defineComponent({
   name: 'NotFound',
