+++ browser/src/components/testing/ImagePredictionResultsFrame.vue
@@ -26,7 +26,7 @@
 </template>
 
 <script>
-import FileEarmarkRuledVue from '../../../assets/svg/FileEarmarkRuled.vue'
+import FileEarmarkRuledVue from '../../assets/svg/FileEarmarkRuled.vue'
 
 export default {
   name: 'ImagePredictionResultsFrame',
