+++ browser/src/components/testing/TestingFrame.vue
@@ -30,10 +30,10 @@
 
 <script>
 import DatasetInputFrame from '../dataset_input/DatasetInputFrame.vue'
-import CustomButton from '../../simple/CustomButton.vue'
-import ModelActorFrame from './ModelActorFrame.vue'
+import CustomButton from '../simple/CustomButton.vue'
+import ModelActorFrame from '../ModelActorFrame.vue'
 
-import { Tester } from '../../../testing/tester'
+import { Tester } from '../../testing/tester'
 import { Task } from 'discojs'
 
 export default {
