+++ browser/src/components/training/TrainingInformationFrame.vue
